Prevent custom export settings from loading when user is not admin
